Move encoding for results into BaseWebResult to prepare for class merging.
